[Show/hide on double click] A double click hides the the children of an item
The mechanism is to pass tree object to qt node item to be used on the double click callback.
The qt node item also receives the uuid and store it for the callback.
The callback will change the item visibility (new variable on visibility.py) and trigger a refresh, redraw.

The visibility now uses str() instead of unique_id.fromMsg to be able to compare the id that is stored on the node_item names

Also triggered the refresh/redraw for the _update_visibility_level()

Known bug: Sometimes there is a segmentation fault after the double click.
It can happens when double click is executed.
Always happens when fit is enabled and the root is double clicked to hide, then it's double clicked to show again.
The segmentation fault happens after the signal redraw slot is executed.